[FE] FIX: 모바일 섹션 스와이프 기능 파일 위치 수정 #796
