+++ frontend/src/pages/Layout.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Outlet } from "react-router";
 import { useNavigate, useLocation } from "react-router-dom";
-import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
+import { useSetRecoilState } from "recoil";
 import { userState } from "@/recoil/atoms";
 import TopNav from "@/components/TopNav/TopNav.container";
 import LeftNav from "@/components/LeftNav/LeftNav";
@@ -12,8 +12,6 @@ import { UserDto } from "@/types/dto/user.dto";
 import styled, { css } from "styled-components";
 import CabinetInfoAreaContainer from "@/components/CabinetInfoArea/CabinetInfoArea.container";
 import MapInfo from "@/components/MapInfo/MapInfo";
-import { currentFloorSectionState } from "@/recoil/selectors";
-import { currentSectionNameState } from "@/recoil/atoms";
 
 const Layout = (): JSX.Element => {
   const [isLoading, setIsLoading] = useState<boolean>(false);
@@ -46,38 +44,6 @@ const Layout = (): JSX.Element => {
     }
   }, []);
 
-  const [touchOffset, setTouchOffset] = useState(0);
-
-  const sectionList = useRecoilValue<Array<string>>(currentFloorSectionState);
-  const [currentSectionName, setCurrentSectionName] = useRecoilState<string>(
-    currentSectionNameState
-  );
-  const currentSectionIdx = sectionList.findIndex(
-    (sectionName) => sectionName === currentSectionName
-  );
-  const moveToLeftSection = () => {
-    if (currentSectionIdx <= 0) {
-      setCurrentSectionName(sectionList[sectionList.length - 1]);
-    } else {
-      setCurrentSectionName(sectionList[currentSectionIdx - 1]);
-    }
-  };
-
-  const moveToRightSection = () => {
-    if (currentSectionIdx >= sectionList.length - 1) {
-      setCurrentSectionName(sectionList[0]);
-    } else {
-      setCurrentSectionName(sectionList[currentSectionIdx + 1]);
-    }
-  };
-
-  const swipeSection = (last: number) => {
-    const swipeDistance = Math.round(last - touchOffset);
-    if (Math.abs(swipeDistance) < 50) return;
-    if (last > touchOffset) moveToLeftSection();
-    else moveToRightSection();
-  };
-
   return isLoginPage ? (
     <Outlet />
   ) : (
@@ -88,14 +54,7 @@ const Layout = (): JSX.Element => {
       ) : (
         <WrapperStyled>
           <LeftNav isVisible={!isHomePage} />
-          <MainStyled
-            onTouchStart={(e: React.TouchEvent) =>
-              setTouchOffset(e.changedTouches[0].screenX)
-            }
-            onTouchEnd={(e: React.TouchEvent) =>
-              swipeSection(e.changedTouches[0].screenX)
-            }
-          >
+          <MainStyled>
             <Outlet />
           </MainStyled>
           <DetailInfoContainerStyled
